Fix analytics page static export
